whisper_integration/requirements.txt
# Core dependencies
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
soundfile>=0.12.1
librosa>=0.10.0
pydub>=0.25.1
PyYAML>=6.0.0
tqdm>=4.65.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
